pyproject.toml
[project]
name = "march-madness-2026"
version = "0.1.0"
description = "Local March Madness bracket portfolio engine, research stack, and release tooling"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.10"
dependencies = [
  "numpy>=2.0.2",
  "duckdb>=1.2.2",
  "httpx>=0.28.1",
  "pandas>=2.3.0",
  "pyarrow>=19.0.1",
  "pydantic>=2.11.2",
  "python-dotenv>=1.0.1",
  "pyyaml>=6.0.2",
  "tenacity>=9.0.0",
  "scipy>=1.14.0",
  "beautifulsoup4>=4.12.0",
  "lxml>=5.0.0",
  "rich>=13.9.0",
  "pyreadr>=0.5.4",
]

[project.optional-dependencies]
gpu = ["cupy-cuda12x>=13.0.0"]

[tool.uv]
package = false
